# sys-service/api/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Aevons 系统服务接口文档，提供用户、角色、部门、岗位、菜单、字典、语言、配置、通知公告和终端应用等系统管理能力",
        "title": "Aevons System Service API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:10702",
    "basePath": "/api/v1/sys",
    "paths": {
...
        "/api/v1/sys/menu": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "新增菜单",
                "parameters": [
                    {
                        "description": "新增菜单",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateMenuDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/menu/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "按语言查询菜单列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/menu/{ids}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "批量删除菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "逗号分隔的菜单ID",
                        "name": "ids",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/menu/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "获取菜单详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "修改菜单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改菜单",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMenuDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/notice": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通知公告"
                ],
                "summary": "新增通知公告",
                "parameters": [
                    {
                        "description": "新增通知公告",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateNoticeDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/notice/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通知公告"
                ],
                "summary": "查询通知公告列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/notice/page": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通知公告"
                ],
                "summary": "分页查询通知公告",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/notice/{ids}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通知公告"
                ],
                "summary": "批量删除通知公告",
                "parameters": [
                    {
                        "type": "string",
                        "description": "逗号分隔的通知公告ID",
                        "name": "ids",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/notice/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通知公告"
                ],
                "summary": "获取通知公告详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "通知公告ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通知公告"
                ],
                "summary": "修改通知公告",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "通知公告ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改通知公告",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateNoticeDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateMenuDTO": {
            "type": "object",
            "required": [
                "isFrame",
                "parentId",
                "sort",
                "status",
                "translations",
                "type",
                "visible"
            ],
            "properties": {
                "activeId": {
                    "description": "激活ID",
                    "type": "string",
                    "example": "0"
                },
                "component": {
                    "description": "组件路径",
                    "type": "string",
                    "maxLength": 100
                },
                "icon": {
                    "description": "图标",
                    "type": "string",
                    "maxLength": 64
                },
                "isFrame": {
                    "description": "是否为外链（0否 1是）",
                    "type": "integer"
                },
                "parentId": {
                    "description": "父级菜单Id",
                    "type": "string",
                    "example": "0"
                },
                "path": {
                    "description": "路由地址",
                    "type": "string",
                    "maxLength": 100
                },
                "permission": {
                    "description": "权限标识",
                    "type": "string",
                    "maxLength": 32
                },
                "query": {
                    "description": "路由参数",
                    "type": "string",
                    "maxLength": 255
                },
                "sort": {
                    "description": "顺序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态（0正常 1停用）",
                    "type": "integer"
                },
                "translations": {
                    "description": "语言信息",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "type": {
                    "description": "类型（1目录 2菜单 3按钮）",
                    "type": "integer"
                },
                "visible": {
                    "description": "是否可见（0隐藏 1显示）",
                    "type": "integer"
                }
            }
        },
        "dto.CreateNoticeDTO": {
            "type": "object",
            "required": [
                "content",
                "status",
                "title",
                "type"
            ],
            "properties": {
                "content": {
                    "description": "公告内容",
                    "type": "string",
                    "maxLength": 3000
                },
                "remark": {
                    "description": "备注",
                    "type": "string",
                    "maxLength": 255
                },
                "status": {
                    "description": "状态（0正常 1关闭）",
                    "type": "integer"
                },
                "title": {
                    "description": "公告标题",
                    "type": "string",
                    "maxLength": 50
                },
                "type": {
                    "description": "公告类型（1通知 2公告）",
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateMenuDTO": {
            "type": "object",
            "required": [
                "id",
                "isFrame",
                "parentId",
                "sort",
                "status",
                "translations",
                "type",
                "visible"
            ],
            "properties": {
                "activeId": {
                    "description": "激活ID",
                    "type": "string",
                    "example": "0"
                },
                "component": {
                    "description": "组件路径",
                    "type": "string",
                    "maxLength": 100
                },
                "icon": {
                    "description": "图标",
                    "type": "string",
                    "maxLength": 64
                },
                "id": {
                    "description": "菜单ID",
                    "type": "string",
                    "example": "0"
                },
                "isFrame": {
                    "description": "是否为外链（0否 1是）",
                    "type": "integer"
                },
                "parentId": {
                    "description": "父级菜单Id",
                    "type": "string",
                    "example": "0"
                },
                "path": {
                    "description": "路由地址",
                    "type": "string",
                    "maxLength": 100
                },
                "permission": {
                    "description": "权限标识",
                    "type": "string",
                    "maxLength": 32
                },
                "query": {
                    "description": "路由参数",
                    "type": "string",
                    "maxLength": 255
                },
                "sort": {
                    "description": "顺序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态（0正常 1停用）",
                    "type": "integer"
                },
                "translations": {
                    "description": "语言信息",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "type": {
                    "description": "类型（1目录 2菜单 3按钮）",
                    "type": "integer"
                },
                "visible": {
                    "description": "是否可见（0隐藏 1显示）",
                    "type": "integer"
                }
            }
        },
        "dto.UpdateNoticeDTO": {
            "type": "object",
            "required": [
                "content",
                "id",
                "status",
                "title",
                "type"
            ],
            "properties": {
                "content": {
                    "description": "公告内容",
                    "type": "string",
                    "maxLength": 3000
                },
                "id": {
                    "description": "公告ID",
                    "type": "integer"
                },
                "remark": {
                    "description": "备注",
                    "type": "string",
                    "maxLength": 255
                },
                "status": {
                    "description": "状态（0正常 1关闭）",
                    "type": "integer"
                },
                "title": {
                    "description": "公告标题",
                    "type": "string",
                    "maxLength": 50
                },
                "type": {
                    "description": "公告类型（1通知 2公告）",
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer 访问令牌，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}